requirements.txt
instructor==1.7.9
python-dotenv==1.0.1
openai==1.70.0
pydantic==2.11.2
